+++ modules/frontend/config.go
@@ -10,6 +10,7 @@ import (
 	"github.com/grafana/tempo/modules/frontend/pipeline"
 	v1 "github.com/grafana/tempo/modules/frontend/v1"
 	"github.com/grafana/tempo/pkg/usagestats"
+	"github.com/grafana/tempo/pkg/util"
 )
 
 var statVersion = usagestats.NewString("frontend_version")
@@ -80,7 +81,7 @@ type SLOConfig struct {
 	ThroughputBytesSLO float64       `yaml:"throughput_bytes_slo,omitempty"`
 }
 
-func (cfg *Config) RegisterFlagsAndApplyDefaults(string, *flag.FlagSet) {
+func (cfg *Config) RegisterFlagsAndApplyDefaults(prefix string, f *flag.FlagSet) {
 	slo := SLOConfig{
 		DurationSLO:        0,
 		ThroughputBytesSLO: 0,
@@ -130,17 +131,15 @@ func (cfg *Config) RegisterFlagsAndApplyDefaults(string, *flag.FlagSet) {
 		MaxTraceQLConditions: 4,
 	}
 
-	// enabling an mcp server opens the door to send tracing data to an LLM. it should require
-	// explicit enabling
-	cfg.MCPServer = MCPServerConfig{
-		Enabled: false,
-	}
-
 	// set default max query size to 128 KiB, queries larger than this will be rejected
 	cfg.MaxQueryExpressionSizeBytes = 128 * 1024
 	// enable multi tenant queries by default
 	cfg.MultiTenantQueriesEnabled = true
 	cfg.Metrics.MaxIntervals = 10_000
+
+	// enabling an mcp server opens the door to send tracing data to an LLM. it should require
+	// explicit enabling. registers a flag in addition to YAML configuration.
+	f.BoolVar(&cfg.MCPServer.Enabled, util.PrefixConfig(prefix, "mcp-server.enabled"), false, "Set to true to enable the MCP server")
 }
 
 type CortexNoQuerierLimits struct{}